战队管理完成

+++ gaming-manage-system/src/main/java/com/ygccw/wechat/zone/service/MatchZoneModelService.java
@@ -1,11 +1,18 @@
 package com.ygccw.wechat.zone.service;
 
+import com.ygccw.wechat.common.recommend.entity.Recommend;
+import com.ygccw.wechat.common.recommend.entity.RecommendMapping;
+import com.ygccw.wechat.common.recommend.enums.RecommendType;
+import com.ygccw.wechat.common.recommend.service.RecommendMappingService;
+import com.ygccw.wechat.common.recommend.service.RecommendService;
 import com.ygccw.wechat.common.zone.entity.MatchZone;
 import com.ygccw.wechat.common.zone.entity.MatchZoneBonus;
 import com.ygccw.wechat.common.zone.entity.MatchZoneCalendar;
 import com.ygccw.wechat.common.zone.service.MatchZoneBonusService;
 import com.ygccw.wechat.common.zone.service.MatchZoneCalendarService;
 import com.ygccw.wechat.common.zone.service.MatchZoneService;
+import com.ygccw.wechat.recommend.model.RecommendMappingModel;
+import com.ygccw.wechat.recommend.service.RecommendMappingModelService;
 import com.ygccw.wechat.zone.model.MatchZoneModel;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
@@ -26,6 +33,12 @@ public class MatchZoneModelService {
     private MatchZoneBonusService matchZoneBonusService;
     @Inject
     private MatchZoneCalendarService matchZoneCalendarService;
+    @Inject
+    private RecommendMappingModelService recommendMappingModelService;
+    @Inject
+    private RecommendService recommendService;
+    @Inject
+    private RecommendMappingService recommendMappingService;
 
 
     @Transactional
@@ -33,13 +46,27 @@ public class MatchZoneModelService {
         MatchZone matchZone = new MatchZone();
         BeanUtils.copyProperties(matchZoneModel, matchZone);
         matchZoneService.save(matchZone);
-        for (MatchZoneBonus matchZoneBonus : matchZoneModel.getMatchZoneBonusList()) {
-            matchZoneBonus.setMatchZoneId(matchZone.getId());
-            matchZoneBonusService.save(matchZoneBonus);
+        if (matchZoneModel.getMatchZoneBonusList() != null) {
+            for (MatchZoneBonus matchZoneBonus : matchZoneModel.getMatchZoneBonusList()) {
+                matchZoneBonus.setMatchZoneId(matchZone.getId());
+                matchZoneBonusService.save(matchZoneBonus);
+            }
         }
-        for (MatchZoneCalendar matchZoneCalendar : matchZoneModel.getMatchZoneCalendarList()) {
-            matchZoneCalendar.setMatchZoneId(matchZone.getId());
-            matchZoneCalendarService.save(matchZoneCalendar);
+        if (matchZoneModel.getMatchZoneCalendarList() != null) {
+            for (MatchZoneCalendar matchZoneCalendar : matchZoneModel.getMatchZoneCalendarList()) {
+                matchZoneCalendar.setMatchZoneId(matchZone.getId());
+                matchZoneCalendarService.save(matchZoneCalendar);
+            }
+        }
+        if (matchZoneModel.getRecommendMappingModelList() != null) {
+            for (RecommendMappingModel recommendMappingModel : matchZoneModel.getRecommendMappingModelList()) {
+                if (recommendMappingModel.getChecked()) {
+                    RecommendMapping recommendMapping = new RecommendMapping();
+                    BeanUtils.copyProperties(recommendMappingModel, recommendMapping);
+                    recommendMapping.setEntityId(matchZone.getId());
+                    recommendMappingService.save(recommendMapping);
+                }
+            }
         }
     }
 
@@ -49,16 +76,37 @@ public class MatchZoneModelService {
         BeanUtils.copyProperties(matchZoneModel, matchZone);
         matchZoneService.update(matchZone);
         matchZoneBonusService.deleteByMatchZoneId(matchZone.getId());
-        for (MatchZoneBonus matchZoneBonus : matchZoneModel.getMatchZoneBonusList()) {
-            matchZoneBonus.setMatchZoneId(matchZone.getId());
-            matchZoneBonus.setId(null);
-            matchZoneBonusService.save(matchZoneBonus);
+        if (matchZoneModel.getMatchZoneBonusList() != null) {
+            for (MatchZoneBonus matchZoneBonus : matchZoneModel.getMatchZoneBonusList()) {
+                matchZoneBonus.setMatchZoneId(matchZone.getId());
+                matchZoneBonus.setId(null);
+                matchZoneBonusService.save(matchZoneBonus);
+            }
         }
-        matchZoneCalendarService.deleteByMatchZoneId(matchZone.getId());
-        for (MatchZoneCalendar matchZoneCalendar : matchZoneModel.getMatchZoneCalendarList()) {
-            matchZoneCalendar.setMatchZoneId(matchZone.getId());
-            matchZoneCalendar.setId(null);
-            matchZoneCalendarService.save(matchZoneCalendar);
+        if (matchZoneModel.getMatchZoneCalendarList() != null) {
+            matchZoneCalendarService.deleteByMatchZoneId(matchZone.getId());
+            for (MatchZoneCalendar matchZoneCalendar : matchZoneModel.getMatchZoneCalendarList()) {
+                matchZoneCalendar.setMatchZoneId(matchZone.getId());
+                matchZoneCalendar.setId(null);
+                matchZoneCalendarService.save(matchZoneCalendar);
+            }
+        }
+        if (matchZoneModel.getRecommendMappingModelList() != null) {
+            for (RecommendMappingModel recommendMappingModel : matchZoneModel.getRecommendMappingModelList()) {
+                RecommendMapping recommendMapping = new RecommendMapping();
+                BeanUtils.copyProperties(recommendMappingModel, recommendMapping);
+                if (recommendMappingModel.getChecked()) {
+                    if (recommendMapping.getId() == null) {
+                        recommendMappingService.save(recommendMapping);
+                    } else {
+                        recommendMappingService.update(recommendMapping);
+                    }
+                } else {
+                    if (recommendMapping.getId() != null) {
+                        recommendMappingService.delete(recommendMapping.getId());
+                    }
+                }
+            }
         }
     }
 
@@ -81,6 +129,25 @@ public class MatchZoneModelService {
         BeanUtils.copyProperties(matchZone, matchZoneModel);
         matchZoneModel.setMatchZoneBonusList(matchZoneBonusService.listByMatchZoneId(matchZone.getId()));
         matchZoneModel.setMatchZoneCalendarList(matchZoneCalendarService.listByMatchZoneId(matchZone.getId()));
+        List<Recommend> recommendList = recommendService.listByType(RecommendType.matchZone);
+        List<RecommendMappingModel> recommendMappingModelList = recommendMappingModelService.listByEntityAndTyp(id, RecommendType.matchZone);
+        List<RecommendMappingModel> recommendMappingModelListNew = new ArrayList<>();
+        for (Recommend recommend : recommendList) {
+            RecommendMappingModel recommendMappingModelNew = new RecommendMappingModel();
+            recommendMappingModelNew.setChecked(false);
+            recommendMappingModelNew.setRecommendName(recommend.getName());
+            recommendMappingModelNew.setEntityId(id);
+            recommendMappingModelNew.setRecommendId(recommend.getId());
+            recommendMappingModelNew.setRecommendType(RecommendType.matchZone);
+            for (RecommendMappingModel recommendMappingModel : recommendMappingModelList) {
+                if (recommendMappingModel.getRecommendId().compareTo(recommend.getId()) == 0) {
+                    BeanUtils.copyProperties(recommendMappingModel, recommendMappingModelNew);
+                    recommendMappingModelNew.setChecked(true);
+                }
+            }
+            recommendMappingModelListNew.add(recommendMappingModelNew);
+        }
+        matchZoneModel.setRecommendMappingModelList(recommendMappingModelListNew);
         return matchZoneModel;
 
     }
